+++ docs/changelog.md
@@ -86,6 +86,8 @@ As features stabilize some brief notes about them will accumulate here.
 * imgcat broken when using `wezterm ssh`. #3624
 * macOS: pasting file(s) will now paste the quoted full path(s) rather than
   just the basename of the path. Thanks to @junnplus! #3663 #3665
+* WebGpu: don't panic when trying to allocate a texture larger than is
+  supported by the GPU. #3713
 
 #### Updated
 * Bundled harfbuzz to 7.1.0